users/to1ne: Lower TAPPING_TERM to 125
By default the TAPPING_TERM is 200ms, this is way too much for me. Lower
it.

The reason it's too much is (I think) because HOLD_ON_OTHER_KEY_PRESS
[1] is not enabled by default (and I think it was back in the day).
While it helped me back in the day to deal with a longer TAPPING_TERM,
it also caused quite some false positives. With this setting disabled
and TAPPING_TERM at 125ms I'm not having a lot of false tap-hold
keypresses.

I should have followed the guidelines[2] from the start:

    For that reason, HOLD_ON_OTHER_KEY_PRESS is strongly discouraged for
    home row mods.

[1]: https://docs.qmk.fm/tap_hold#hold-on-other-key-press
[2]: https://precondition.github.io/home-row-mods#hold-on-other-key-press